Separate TensorWrapper class. * On going work to make SWIG happy.
